chore(utils): move dependencies to utils root
